Type editor suggestions
